Avoid needing to allocate a secrets.Bag
